# bezier: optimize setup of bezier curves