avoid reallocations by resizing and copy the pVec into the resized m_dPoints, reduce the amount of calculations in baking to only do it once per iteration instead of twice. precompute in getYforT and getXforT return early in getYForPoint if x is equal or below 0. and use const references where we can.

these changes we are now down to an average of "time to bake: 2.50µs."
on my machine compared to before average of "time to bake: 11.15µs"

patched
```
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.92µs. Estimated average calc time: 0.10µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.65µs. Estimated average calc time: 0.07µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.53µs. Estimated average calc time: 0.07µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.48µs. Estimated average calc time: 0.08µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.57µs. Estimated average calc time: 0.04µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.33µs. Estimated average calc time: 0.07µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.36µs. Estimated average calc time: 0.02µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.36µs. Estimated average calc time: 0.04µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.52µs. Estimated average calc time: 0.07µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.33µs. Estimated average calc time: 0.07µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.45µs. Estimated average calc time: 0.06µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.34µs. Estimated average calc time: 0.06µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.34µs. Estimated average calc time: 0.05µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 2.36µs. Estimated average calc time: 0.05µs.
```

and unpatched

```
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 46.05µs. Estimated average calc time: 0.09µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 11.60µs. Estimated average calc time: 0.05µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 11.24µs. Estimated average calc time: 0.06µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 11.18µs. Estimated average calc time: 0.06µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 11.19µs. Estimated average calc time: 0.03µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 11.20µs. Estimated average calc time: 0.06µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 11.22µs. Estimated average calc time: 0.02µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 11.14µs. Estimated average calc time: 0.05µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 11.17µs. Estimated average calc time: 0.07µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 11.14µs. Estimated average calc time: 0.06µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 11.20µs. Estimated average calc time: 0.05µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 11.18µs. Estimated average calc time: 0.05µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 10.65µs. Estimated average calc time: 0.05µs.
[LOG] Created a bezier curve, baked 255 points, mem usage: 4.08kB, time to bake: 13.77µs. Estimated average calc time: 0.08µs.
```