@gulafaran 2024-11-21
> @gulafaran, is it worthy to replace `float` with `float const&`? Aren't they both 4 bytes, and you get to skip one indirection by not using references? (The compiler probably optimizes that out, but then why the change?)

yeah thats true, it slinged out in mere force of habit. it probably is faster with just passing by value and not having a potential indirections, atleast when it comes to small types like this. float/int etc. now if it even makes an measurable difference i doubt it heh