@gulafaran 2024-11-21
> > @gulafaran, is it worthy to replace `float` with `float const&`? Aren't they both 4 bytes, and you get to skip one indirection by not using references? (The compiler probably optimizes that out, but then why the change?)
> 
> yeah thats true, it slinged out in mere force of habit. it probably is faster with just passing by value and not having a potential indirections, atleast when it comes to small types like this. float/int etc. now if it even makes an measurable difference i doubt it heh

also from what i understand it also entirerly depends on your architecture, compiler, the phase shift of the moon and your bitbots aligning in your cache to tell what is faster and what is not when it comes to these smaller types :D so il leave the decision to vaxry how he wants it